fix: Deleting Purchase Method breaks Products

+++ repos/fdbt-site/src/pages/api/deletePurchaseMethod.ts
@@ -1,7 +1,14 @@
 import { NextApiResponse } from 'next';
-import { deleteSalesOfferPackageByNocCodeAndName } from '../../data/auroradb';
+import {
+    deleteSalesOfferPackageByNocCodeAndName,
+    getAllProductsByNoc,
+    getSalesOfferPackageByIdAndNoc,
+} from '../../data/auroradb';
 import { redirectToError, redirectTo, getAndValidateNoc } from '../../utils/apiUtils/index';
-import { NextApiRequestWithSession } from '../../interfaces';
+import { ErrorInfo, NextApiRequestWithSession } from '../../interfaces';
+import { getProductsMatchingJson } from 'src/data/s3';
+import { updateSessionAttribute } from 'src/utils/sessions';
+import { VIEW_PURCHASE_METHOD } from 'src/constants/attributes';
 
 export default async (req: NextApiRequestWithSession, res: NextApiResponse): Promise<void> => {
     try {
@@ -12,6 +19,34 @@ export default async (req: NextApiRequestWithSession, res: NextApiResponse): Pro
             throw new Error('insufficient data provided for delete query');
         }
         const nocCode = getAndValidateNoc(req, res);
+
+        const products = await getAllProductsByNoc(nocCode);
+        const matchingJsonLinks = products.map((product) => product.matchingJsonLink);
+        const tickets = await Promise.all(
+            matchingJsonLinks.map(async (link) => {
+                return await getProductsMatchingJson(link);
+            }),
+        );
+
+        let productsUsingPurchaseMethod = [];
+        for (const ticket of tickets) {
+            for (const salesOfferPackage of ticket.products[0].salesOfferPackages) {
+                if (salesOfferPackage.id === id) {
+                    productsUsingPurchaseMethod.push(ticket);
+                }
+            }
+        }
+
+        if (productsUsingPurchaseMethod.length > 0) {
+            const purchaseDetails = await getSalesOfferPackageByIdAndNoc(id, nocCode);
+            const { name } = purchaseDetails;
+            const errorMessage = `You cannot delete ${name} because it is being used in ${productsUsingPurchaseMethod.length} products.`;
+            const errors: ErrorInfo[] = [{ id: '/viewPurchaseMethods', errorMessage }];
+            updateSessionAttribute(req, VIEW_PURCHASE_METHOD, errors);
+            redirectTo(res, `/viewPurchaseMethods?cannotDelete=${name}`);
+            return;
+        }
+
         await deleteSalesOfferPackageByNocCodeAndName(id, nocCode);
 
         redirectTo(res, '/viewPurchaseMethods');
